frontend: Move `services.ts` file to `/src/common/services`
