+++ frontend/src/common/components/mutateComponents/OwnerMutateForm.tsx
@@ -3,8 +3,8 @@ import {Button, IconArrowLeft} from "hds-react";
 import {useEffect, useState} from "react";
 import {useForm} from "react-hook-form";
 import {z} from "zod";
-import {useSaveOwnerMutation} from "../../../app/services";
 import {IOwner, OwnerSchema} from "../../schemas";
+import {useSaveOwnerMutation} from "../../services";
 import {hdsToast, validateBusinessId, validateSocialSecurityNumber} from "../../utils";
 import {CheckboxInput, TextInput} from "../forms";
 import FormProviderForm from "../forms/FormProviderForm";
