+++ frontend/src/features/apartment/ApartmentListPage.tsx
@@ -2,7 +2,6 @@ import React, {useState} from "react";
 
 import {IconSearch, LoadingSpinner, StatusLabel} from "hds-react";
 import {Link} from "react-router-dom";
-import {useGetApartmentsQuery, useGetHousingCompanyApartmentsQuery} from "../../app/services";
 import {Heading, ListPageNumbers, QueryStateHandler} from "../../common/components";
 import {
     FilterCheckboxField,
@@ -12,6 +11,7 @@ import {
 } from "../../common/components/filters";
 import {getApartmentSoldStatusLabel} from "../../common/localisation";
 import {IApartment, IApartmentListResponse} from "../../common/schemas";
+import {useGetApartmentsQuery, useGetHousingCompanyApartmentsQuery} from "../../common/services";
 import {formatDate} from "../../common/utils";
 import ConditionsOfSaleStatus from "./components/ConditionsOfSaleStatus";
