+++ frontend/src/features/apartment/ApartmentNewSalePage/ApartmentSalePage.tsx
@@ -5,10 +5,10 @@ import {useForm} from "react-hook-form";
 import {useNavigate} from "react-router-dom";
 import {v4 as uuidv4} from "uuid";
 import {z, ZodSchema} from "zod";
-import {useCreateSaleMutation} from "../../../app/services";
 import {ConfirmDialogModal, Heading, NavigateBackButton, OwnershipList, SaveButton} from "../../../common/components";
 import {FormProviderForm} from "../../../common/components/forms";
 import {IApartmentSaleForm, OwnershipsListSchema} from "../../../common/schemas";
+import {useCreateSaleMutation} from "../../../common/services";
 import {hdsToast, setAPIErrorsForFormFields, today} from "../../../common/utils";
 import {ApartmentViewContext, ApartmentViewContextProvider} from "../components/ApartmentViewContextProvider";
 import {ApartmentCatalogPricesFieldSet, ApartmentSaleFormFieldSet, MaximumPriceCalculationFieldSet} from "./fieldsets";
