+++ frontend/src/features/codes/CodesPage.tsx
@@ -1,6 +1,5 @@
 import {Tabs} from "hds-react";
 import React from "react";
-import {useGetDevelopersQuery, useGetOwnersQuery, useGetPropertyManagersQuery} from "../../app/services";
 import {
     developerMutateFormProps,
     MutateForm,
@@ -8,6 +7,7 @@ import {
     OwnerMutateForm,
     propertyManagerMutateFormProps,
 } from "../../common/components/mutateComponents";
+import {useGetDevelopersQuery, useGetOwnersQuery, useGetPropertyManagersQuery} from "../../common/services";
 import IndicesList from "./IndicesList";
 
 const CodesPage = (): React.JSX.Element => {
