+++ frontend/src/features/functions/ThirtyYearRegulation.tsx
@@ -3,15 +3,15 @@ import {FormProvider, useForm} from "react-hook-form";
 import {hitasQuarters} from "../../common/schemas";
 
 import {useState} from "react";
+import {Divider, DownloadButton, Heading, QueryStateHandler} from "../../common/components";
+import {SelectInput, ToggleInput} from "../../common/components/forms";
 import {
     downloadRegulationResults,
     useCreateThirtyYearRegulationMutation,
     useGetExternalSalesDataQuery,
     useGetIndexQuery,
     useGetThirtyYearRegulationQuery,
-} from "../../app/services";
-import {Divider, DownloadButton, Heading, QueryStateHandler} from "../../common/components";
-import {SelectInput, ToggleInput} from "../../common/components/forms";
+} from "../../common/services";
 import {getHitasQuarter, hdsToast} from "../../common/utils";
 import {ExternalSalesDataImport, ThirtyYearErrorModal, ThirtyYearErrorTest, ThirtyYearResults} from "./components";
