+++ frontend/src/features/functions/components/ExternalSalesDataImport.tsx
@@ -1,8 +1,8 @@
 import {useState} from "react";
 import {FormProvider, useForm} from "react-hook-form";
-import {useSaveExternalSalesDataMutation} from "../../../app/services";
 import {SaveDialogModal} from "../../../common/components";
 import {FileInput, FormProviderForm} from "../../../common/components/forms";
+import {useSaveExternalSalesDataMutation} from "../../../common/services";
 import {hdsToast} from "../../../common/utils";
 
 export default function ExternalSalesDataImport({formDate}) {
