+++ frontend/src/features/reports/components/OwnerReports.tsx
@@ -1,5 +1,5 @@
-import {downloadMultipleOwnershipsReportPDF} from "../../../app/services";
 import {DownloadButton} from "../../../common/components";
+import {downloadMultipleOwnershipsReportPDF} from "../../../common/services";
 
 const OwnerReports = () => {
     return (
